Put all error codes in one errorcode enum.
